Fix manila test jobs
test code on this branch should not be
testing snapshots nor cloning of snapshots.
Snapshots are disabled by default and
the cloning functionality is not available
in manila's cephfs drivers in this stable
branch.

We can also get rid of the devstack-gate
hooks that are no longer needed since
the jobs are migrated to zuulv3 style.

Depends-On: I22ef492ec84d2c1b864a461f285e2eb2e14bf617
Change-Id: Ie65b6c0877c5cc6976bd4bbba635d7654a90c3cf
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>